板を回転するカスタムフック
