+++ src/components/OrigamiPost/index.tsx
@@ -1,11 +1,10 @@
 "use client";
 
-import React, { useEffect, useRef, useState } from "react";
+import React, { useRef, useState } from "react";
 import styles from "./index.module.scss";
 import * as THREE from "three";
 import { OrbitControls } from "three/examples/jsm/Addons.js";
 import { rotateBoards } from "./logics/rotateBoards";
-import { renderBoard } from "./logics/renderBoard";
 import { Point, Board } from "@/types/three";
 import { FoldMethodControlPanel } from "./FoldMethodControlPanel";
 import { Step } from "./FoldMethodControlPanel";
@@ -20,6 +19,7 @@ import { useSelectPoints } from "./hooks/useSelectPoints";
 import { useDecideRotateAxis } from "./hooks/useDecideRotateAxis";
 import { useDecideTargetBoard } from "./hooks/useDecideTargetBoard";
 import { useSelectSideAndNumberOfBoards } from "./hooks/useSelectSideAndNumberOfBoards";
+import { useRotateBoards } from "./hooks/useRotateBoards";
 
 export const OrigamiPost = () => {
   const initialBoard: Board = [
@@ -122,61 +122,18 @@ export const OrigamiPost = () => {
   });
 
   // 回転に応じて板を描画
-  useEffect(() => {
-    const scene = sceneRef.current;
-    if (!scene) return;
-    if (inputStep !== "fold") return;
-    if (rotateAxis.length === 0) return;
-
-    // TODO: 共通化する
-    // xy平面上の板のうち、z座標が大きい順に、numberOfMoveBoards枚を折る
-    // それ以外の板は無条件で折る
-    let xyPlaneBoards: Board[] = [];
-    const notXyPlaneBoards: Board[] = [];
-    for (let i = 0; i < moveBoards.length; i++) {
-      const board = moveBoards[i];
-      const isEquallyZ = board.every((point) => point[2] === board[0][2]);
-      if (isEquallyZ) {
-        xyPlaneBoards.push(board);
-      } else {
-        notXyPlaneBoards.push(board);
-      }
-    }
-
-    // xy平面上の板をz座標が大きい順にソート
-    xyPlaneBoards = xyPlaneBoards.sort((a, b) => b[0][2] - a[0][2]);
-
-    const foldBoards = [
-      ...xyPlaneBoards.slice(0, numberOfMoveBoards),
-      ...notXyPlaneBoards,
-    ];
-    const notFoldBoards = xyPlaneBoards.slice(numberOfMoveBoards);
-
-    const rotatedBoards = rotateBoards({
-      boards: foldBoards,
-      rotateAxis,
-      angle: foldingAngle,
-      isFoldingDirectionFront: isFoldingDirectionFront,
-      isMoveBoardsRight,
-    });
-    const boards = [...fixBoards, ...rotatedBoards, ...notFoldBoards];
-
-    // 前の板を削除
-    scene.children = scene.children.filter((child) => {
-      return child.name === "Axis";
-    });
-    // 板を描画
-    boards.forEach((board) => {
-      renderBoard({ scene, board, color: origamiColor });
-    });
-  }, [
-    foldingAngle,
-    fixBoards,
-    moveBoards,
+  useRotateBoards({
+    sceneRef,
+    inputStep,
     rotateAxis,
+    foldingAngle,
+    isFoldingDirectionFront,
     numberOfMoveBoards,
+    isMoveBoardsRight,
+    moveBoards,
+    fixBoards,
     origamiColor,
-  ]);
+  });
 
   const handleDecideFoldMethod = () => {
     // moveBoardsを回転した後の板を、fixBoardsに追加する
